refactor: added shared iban data
